renamed enums file to routes
